requirements.txt
requests
beautifulsoup4
google-generativeai
python-dotenv
googlesearch-python
markdown
resend
